// src/assets/js/charting_library/charting_library.min.d.ts
export declare type LanguageCode =
  | 'ar'
  | 'zh'
  | 'cs'
  | 'da_DK'
  | 'nl_NL'
  | 'en'
  | 'et_EE'
  | 'fr'
  | 'de'
  | 'el'
  | 'he_IL'
  | 'hu_HU'
  | 'id_ID'
  | 'it'
  | 'ja'
  | 'ko'
  | 'fa'
  | 'pl'
  | 'pt'
  | 'ro'
  | 'ru'
  | 'sk_SK'
  | 'es'
  | 'sv'
  | 'th'
  | 'tr'
  | 'vi';
// ...
export interface IFormatter {
  format(value: any): string;
  parse?(value: string): ErrorFormatterParseResult | SuccessFormatterParseResult;
}
export declare type StudyInputValueType = string | number | boolean;
export declare type StudyOverrideValueType = string | number | boolean;
export interface StudyOverrides {
  [key: string]: StudyOverrideValueType;
}
// ...
export interface IWatchedValue<T> extends IWatchedValueReadonly<T> {
  value(): T;
  setValue(value: T, forceUpdate?: boolean): void;
  subscribe(callback: WatchedValueCallback<T>, options?: WatchedValueSubscribeOptions): void;
  unsubscribe(callback?: WatchedValueCallback<T> | null): void;
  readonly(): IWatchedValueReadonly<T>;
  spawn(): IWatchedValueSpawn<T>;
}
// ...
export declare type ResolutionString = string;
// ...
export interface IDatafeedQuotesApi {
  getQuotes(symbols: string[], onDataCallback: QuotesCallback, onErrorCallback: (msg: string) => void): void;
  subscribeQuotes(symbols: string[], fastSymbols: string[], onRealtimeCallback: QuotesCallback, listenerGUID: string): void;
  unsubscribeQuotes(listenerGUID: string): void;
}
// ...
export declare type IBasicDataFeed = IDatafeedChartApi & IExternalDatafeed;
export declare type ThemeName = 'Light' | 'Dark';
export interface ChartingLibraryWidgetOptions {
  container_id: string;
  datafeed: IBasicDataFeed | (IBasicDataFeed & IDatafeedQuotesApi);
  interval: ResolutionString;
  symbol: string;
  auto_save_delay?: number;
  autosize?: boolean;
  debug?: boolean;
  disabled_features?: string[];
  drawings_access?: AccessList;
  enabled_features?: string[];
  fullscreen?: boolean;
  height?: number;
  library_path?: string;
  locale: LanguageCode;
  numeric_formatting?: NumericFormattingParams;
  saved_data?: object;
  studies_access?: AccessList;
  study_count_limit?: number;
  symbol_search_request_delay?: number;
  timeframe?: string;
  timezone?: 'exchange' | Timezone;
  toolbar_bg?: string;
  width?: number;
  charts_storage_url?: string;
  charts_storage_api_version?: AvailableSaveloadVersions;
  client_id?: string;
  user_id?: string;
  load_last_chart?: boolean;
  studies_overrides?: StudyOverrides;
  customFormatters?: CustomFormatters;
  overrides?: Overrides;
  snapshot_url?: string;
  indicators_file_name?: string;
  preset?: 'mobile';
  time_frames?: TimeFrameItem[];
  custom_css_url?: string;
  favorites?: Favorites;
  save_load_adapter?: IExternalSaveLoadAdapter;
  loading_screen?: LoadingScreenOptions;
  settings_adapter?: ISettingsAdapter;
  theme?: ThemeName;
}
export interface TradingTerminalWidgetOptions extends ChartingLibraryWidgetOptions {
  brokerConfig?: SingleBrokerMetaInfo;
  restConfig?: RestBrokerMetaInfo;
  widgetbar?: WidgetBarParams;
  rss_news_feed?: RssNewsFeedParams;
  news_provider?: NewsProvider;
  brokerFactory?(host: IBrokerConnectionAdapterHost): IBrokerWithoutRealtime | IBrokerTerminal;
}
export declare type LayoutType = 's' | '2h' | '2-1' | '2v' | '3h' | '3v' | '3s' | '4' | '6' | '8';
export declare type SupportedLineTools =
  | 'text'
  | 'anchored_text'
  | 'note'
  | 'anchored_note'
  | 'double_curve'
  | 'arc'
  | 'icon'
  | 'arrow_up'
  | 'arrow_down'
  | 'arrow_left'
  | 'arrow_right'
  | 'price_label'
  | 'flag'
  | 'vertical_line'
  | 'horizontal_line'
  | 'horizontal_ray'
  | 'trend_line'
  | 'trend_angle'
  | 'arrow'
  | 'ray'
  | 'extended'
  | 'parallel_channel'
  | 'disjoint_angle'
  | 'flat_bottom'
  | 'pitchfork'
  | 'schiff_pitchfork_modified'
  | 'schiff_pitchfork'
  | 'balloon'
  | 'inside_pitchfork'
  | 'pitchfan'
  | 'gannbox'
  | 'gannbox_square'
  | 'gannbox_fixed'
  | 'gannbox_fan'
  | 'fib_retracement'
  | 'fib_trend_ext'
  | 'fib_speed_resist_fan'
  | 'fib_timezone'
  | 'fib_trend_time'
  | 'fib_circles'
  | 'fib_spiral'
  | 'fib_speed_resist_arcs'
  | 'fib_channel'
  | 'xabcd_pattern'
  | 'cypher_pattern'
  | 'abcd_pattern'
  | 'callout'
  | 'triangle_pattern'
  | '3divers_pattern'
  | 'head_and_shoulders'
  | 'fib_wedge'
  | 'elliott_impulse_wave'
  | 'elliott_triangle_wave'
  | 'elliott_triple_combo'
  | 'elliott_correction'
  | 'elliott_double_combo'
  | 'cyclic_lines'
  | 'time_cycles'
  | 'sine_line'
  | 'long_position'
  | 'short_position'
  | 'forecast'
  | 'date_range'
  | 'price_range'
  | 'date_and_price_range'
  | 'bars_pattern'
  | 'ghost_feed'
  | 'projection'
  | 'rectangle'
  | 'rotated_rectangle'
  | 'ellipse'
  | 'triangle'
  | 'polyline'
  | 'curve'
  | 'regression_trend'
  | 'cursor'
  | 'dot'
  | 'arrow_cursor'
  | 'eraser'
  | 'measure'
  | 'zoom'
  | 'brush';
// ...
export interface IChartWidgetApi {
  onDataLoaded(): ISubscription<() => void>;
  onSymbolChanged(): ISubscription<() => void>;
  onIntervalChanged(): ISubscription<
    (
      interval: ResolutionString,
      timeFrameParameters: {
        timeframe?: string;
      }
    ) => void
  >;
  onVisibleRangeChanged(): ISubscription<() => void>;
  dataReady(callback: () => void): boolean;
  crossHairMoved(callback: (params: CrossHairMovedEventParams) => void): void;
  setVisibleRange(range: VisibleTimeRange, callback: () => void): void;
  setSymbol(symbol: string, callback: () => void): void;
  setResolution(resolution: ResolutionString, callback: () => void): void;
  resetData(): void;
  executeActionById(actionId: ChartActionId): void;
  getCheckableActionState(actionId: ChartActionId): boolean;
  refreshMarks(): void;
  clearMarks(): void;
  setChartType(type: SeriesStyle): void;
  getAllShapes(): EntityInfo[];
  getAllStudies(): EntityInfo[];
  /**
   * @deprecated Use shape/study API instead ([getStudyById] / [getShapeById])
   */
  setEntityVisibility(entityId: EntityId, isVisible: boolean): void;
  createStudy<TStudyInputs extends StudyInputValueType, TOverrides extends StudyOverrides>(
    name: string,
    forceOverlay: boolean,
    lock?: boolean,
    inputs?: TStudyInputs[],
    callback?: (entityId: EntityId) => void,
    overrides?: TOverrides,
    options?: CreateStudyOptions
  ): EntityId | null;
  getStudyById(entityId: EntityId): IStudyApi;
  createShape<TOverrides extends object>(point: ShapePoint, options: CreateShapeOptions<TOverrides>): EntityId | null;
  createMultipointShape<TOverrides extends object>(points: ShapePoint[], options: CreateShapeOptions<TOverrides>): EntityId | null;
  getShapeById(entityId: EntityId): ILineDataSourceApi;
  removeEntity(entityId: EntityId): void;
  removeAllShapes(): void;
  removeAllStudies(): void;
  createStudyTemplate(options: CreateStudyTemplateOptions): object;
  applyStudyTemplate(template: object): void;
  createOrderLine(options: CreateTradingPrimitiveOptions): IOrderLineAdapter;
  createPositionLine(options: CreateTradingPrimitiveOptions): IPositionLineAdapter;
  createExecutionShape(options: CreateTradingPrimitiveOptions): IExecutionLineAdapter;
  symbol(): string;
  symbolExt(): SymbolExt;
  resolution(): ResolutionString;
  getVisibleRange(): VisibleTimeRange;
  getVisiblePriceRange(): VisiblePriceRange;
  priceFormatter(): IFormatter;
  chartType(): SeriesStyle;
  setTimezone(timezone: 'exchange' | Timezone): void;
  getPanes(): IPaneApi[];
}
// ...
export declare type EmptyCallback = () => void;
export interface SubscribeEventsMap {
  toggle_sidebar: (isHidden: boolean) => void;
  indicators_dialog: EmptyCallback;
  toggle_header: (isHidden: boolean) => void;
  edit_object_dialog: (params: EditObjectDialogEventParams) => void;
  chart_load_requested: (savedData: object) => void;
  chart_loaded: EmptyCallback;
  mouse_down: (params: MouseEventParams) => void;
  mouse_up: (params: MouseEventParams) => void;
  drawing: (params: StudyOrDrawingAddedToChartEventParams) => void;
  study: (params: StudyOrDrawingAddedToChartEventParams) => void;
  undo: EmptyCallback;
  redo: EmptyCallback;
  reset_scales: EmptyCallback;
  compare_add: EmptyCallback;
  add_compare: EmptyCallback;
  'load_study template': EmptyCallback;
  onTick: (tick: Bar) => void;
  onAutoSaveNeeded: EmptyCallback;
  onScreenshotReady: (url: string) => void;
  onMarkClick: (markId: Mark['id']) => void;
  onTimescaleMarkClick: (markId: TimescaleMark['id']) => void;
  onSelectedLineToolChanged: EmptyCallback;
  layout_about_to_be_changed: (newLayoutType: LayoutType) => void;
  layout_changed: EmptyCallback;
  activeChartChanged: (chartIndex: number) => void;
}
export interface SaveChartToServerOptions {
  chartName?: string;
  defaultChartName?: string;
}
export interface WatchListApi {
  defaultList(): string[];
  getList(id?: string): string[] | null;
  getAllLists(): WatchListSymbolListMap | null;
  getActiveListId(): string | null;
  setList(symbols: string[]): void;
  updateList(listId: string, symbols: string[]): void;
  renameList(listId: string, newName: string): void;
  createList(listName?: string, symbols?: string[]): WatchListSymbolList | null;
  saveList(list: WatchListSymbolList): boolean;
  deleteList(listId: string): void;
  onListChanged(): ISubscription<EmptyCallback>;
  onActiveListChanged(): ISubscription<EmptyCallback>;
  onListAdded(): ISubscription<WatchListSymbolListAddedCallback>;
  onListRemoved(): ISubscription<WatchListSymbolListRemovedCallback>;
  onListRenamed(): ISubscription<WatchListSymbolListRenamedCallback>;
}
export interface GrayedObject {
  type: 'drawing' | 'study';
  name: string;
}
export interface ContextMenuItem {
  position: 'top' | 'bottom';
  text: string;
  click: EmptyCallback;
}
export interface DialogParams<CallbackType> {
  title: string;
  body: string;
  callback: CallbackType;
}
export interface SymbolIntervalResult {
  symbol: string;
  interval: ResolutionString;
}
export interface SaveLoadChartRecord {
  id: string;
  name: string;
  image_url: string;
  modified_iso: number;
  short_symbol: string;
  interval: ResolutionString;
}
export interface CreateButtonOptions {
  align: 'right' | 'left';
}
export interface IChartingLibraryWidget {
  onChartReady(callback: EmptyCallback): void;
  onGrayedObjectClicked(callback: (obj: GrayedObject) => void): void;
  onShortcut(shortCut: string, callback: EmptyCallback): void;
  subscribe<EventName extends keyof SubscribeEventsMap>(event: EventName, callback: SubscribeEventsMap[EventName]): void;
  unsubscribe<EventName extends keyof SubscribeEventsMap>(event: EventName, callback: SubscribeEventsMap[EventName]): void;
  chart(index?: number): IChartWidgetApi;
  setLanguage(lang: LanguageCode): void;
  setSymbol(symbol: string, interval: ResolutionString, callback: EmptyCallback): void;
  remove(): void;
  closePopupsAndDialogs(): void;
  selectLineTool(linetool: SupportedLineTools): void;
  selectedLineTool(): SupportedLineTools;
  save(callback: (state: object) => void): void;
  load(state: object): void;
  getSavedCharts(callback: (chartRecords: SaveLoadChartRecord[]) => void): void;
  loadChartFromServer(chartRecord: SaveLoadChartRecord): void;
  saveChartToServer(onComplete?: EmptyCallback, onFail?: EmptyCallback, saveAsSnapshot?: false, options?: SaveChartToServerOptions): void;
  removeChartFromServer(chartId: string, onCompleteCallback: EmptyCallback): void;
  onContextMenu(callback: (unixTime: number, price: number) => ContextMenuItem[]): void;
  createButton(options?: CreateButtonOptions): JQuery;
  showNoticeDialog(params: DialogParams<() => void>): void;
  showConfirmDialog(params: DialogParams<(confirmed: boolean) => void>): void;
  showLoadChartDialog(): void;
  showSaveAsChartDialog(): void;
  symbolInterval(): SymbolIntervalResult;
  mainSeriesPriceFormatter(): IFormatter;
  getIntervals(): string[];
  getStudiesList(): string[];
  addCustomCSSFile(url: string): void;
  applyOverrides<TOverrides extends StudyOverrides>(overrides: TOverrides): void;
  applyStudiesOverrides(overrides: object): void;
  watchList(): WatchListApi;
  activeChart(): IChartWidgetApi;
  chartsCount(): number;
  layout(): LayoutType;
  setLayout(layout: LayoutType): void;
  changeTheme(themeName: ThemeName): void;
  takeScreenshot(): void;
  lockAllDrawingTools(): IWatchedValue<boolean>;
}
export interface ChartingLibraryWidgetConstructor {
  new (options: ChartingLibraryWidgetOptions | TradingTerminalWidgetOptions): IChartingLibraryWidget;
}
export declare function version(): string;
export declare function onready(callback: () => void): void;
export declare const widget: ChartingLibraryWidgetConstructor;